Moved template instantiation into Edge
